extract component + review 1
