+++ front/components/assistant/conversation/InputBar.tsx
@@ -1,16 +1,5 @@
-import {
-  Avatar,
-  Button,
-  DropdownMenu,
-  IconButton,
-  MagnifyingGlassStrokeIcon,
-  PaperAirplaneIcon,
-  PlusIcon,
-  RobotIcon,
-  WrenchIcon,
-} from "@dust-tt/sparkle";
+import { Avatar, IconButton, PaperAirplaneIcon } from "@dust-tt/sparkle";
 import { Transition } from "@headlessui/react";
-import Link from "next/link";
 import {
   createContext,
   ForwardedRef,
@@ -30,6 +19,7 @@ import { classNames } from "@app/lib/utils";
 import { AgentConfigurationType } from "@app/types/assistant/agent";
 import { MentionType } from "@app/types/assistant/conversation";
 import { WorkspaceType } from "@app/types/user";
+import { AssistantPicker } from "@app/components/assistant/AssistantPicker";
 
 // AGENT MENTION
 
@@ -634,82 +624,6 @@ export function AssistantInputBar({
         </div>
       </div>
     </>
-  );
-}
-
-export function AssistantPicker({
-  owner,
-  assistants,
-  onItemClick,
-}: {
-  owner: WorkspaceType;
-  assistants: AgentConfigurationType[];
-  onItemClick: (assistant: AgentConfigurationType) => void;
-}) {
-  const [searchText, setSearchText] = useState("");
-
-  const searchedAssistants = assistants.filter((a) =>
-    a.name.toLowerCase().startsWith(searchText.toLowerCase())
-  );
-
-  return (
-    <div onBlur={() => setSearchText("")}>
-      <DropdownMenu>
-        <DropdownMenu.Button icon={RobotIcon} />
-        <DropdownMenu.Items origin="bottomRight" width={240}>
-          <div className="border-b border-structure-100 p-2">
-            <div className="relative text-sm font-medium text-element-800">
-              <input
-                type="text"
-                placeholder="Search"
-                value={searchText}
-                className="w-full rounded-full border-structure-200 pr-8"
-                onKeyUp={(e) => {
-                  if (e.key === "Enter" && searchedAssistants.length > 0) {
-                    onItemClick(searchedAssistants[0]);
-                    setSearchText("");
-                  }
-                }}
-                onChange={(e) => {
-                  setSearchText(e.target.value.toLowerCase());
-                }}
-              />
-              <MagnifyingGlassStrokeIcon className="absolute right-3 top-2.5 h-5 w-5" />
-            </div>
-          </div>
-          <div className="max-h-64 overflow-y-auto [&>*]:w-full">
-            {searchedAssistants.map((c) => (
-              <DropdownMenu.Item
-                key={c.sId}
-                label={"@" + c.name}
-                visual={c.pictureUrl}
-                onClick={() => onItemClick(c)}
-              />
-            ))}
-          </div>
-          {(owner.role === "admin" || owner.role === "builder") && (
-            <div className="flex flex-row justify-between border-t border-structure-100 px-3 py-2">
-              <Link href={`/w/${owner.sId}/builder/assistants/new`}>
-                <Button
-                  label="Create"
-                  size="xs"
-                  variant="secondary"
-                  icon={PlusIcon}
-                />
-              </Link>
-              <Link href={`/w/${owner.sId}/builder/assistants`}>
-                <Button
-                  label="Manage"
-                  size="xs"
-                  variant="tertiary"
-                  icon={WrenchIcon}
-                />
-              </Link>
-            </div>
-          )}
-        </DropdownMenu.Items>
-      </DropdownMenu>
-    </div>
   );
 }
